refactor: split feeding, sound, and water pages into smaller components and add responsive layouts
