+++ app/water/page.tsx
@@ -1,21 +1,9 @@
 'use client'
 
-import { FaPlus } from 'react-icons/fa'
 import Image from 'next/image'
 
-import CustomCard from '@/components/CustomCard'
+import WaterPreferences from '@/components/water/WaterPreferences'
 import NavBar from '@/components/NavBar'
-import { Button } from '@/components/ui/button'
-import { CardContent } from '@/components/ui/card'
-import {
-  Drawer,
-  DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerTitle,
-  DrawerTrigger,
-} from '@/components/ui/drawer'
 
 export default function WaterPage() {
   const handleClick = () => {
@@ -34,48 +22,7 @@ export default function WaterPage() {
           onClick={handleClick}
         />
       </div>
-      <Drawer>
-        <DrawerTrigger asChild>
-          <Button className="bg-transparent h-24 shadow-none">
-            Open Drawer
-          </Button>
-        </DrawerTrigger>
-        <DrawerContent>
-          <div className="mx-auto w-full max-w-sm">
-            <DrawerHeader>
-              <DrawerTitle>Water Settings</DrawerTitle>
-              <DrawerDescription>Set your water preferences.</DrawerDescription>
-            </DrawerHeader>
-            <div className="flex flex-col p-4 gap-5">
-              <CustomCard cardTitle="Daily Morning">
-                <CardContent>
-                  <div className="mb-4 grid grid-cols-[25px_1fr] items-start pb-4 last:mb-0 last:pb-0">
-                    <span className="flex h-2 w-2 translate-y-1 rounded-full bg[#F7BE7A]" />
-                    <div className="space-y-2">
-                      <p className="leading-none">9:00 AM, Daily, 100 ml</p>
-                    </div>
-                  </div>
-                </CardContent>
-              </CustomCard>
-              <CustomCard cardTitle="Afternoon">
-                <CardContent>
-                  <div className="mb-4 grid grid-cols-[25px_1fr] items-start pb-4 last:mb-0 last:pb-0">
-                    <span className="flex h-2 w-2 translate-y-1 rounded-full bg[#F7BE7A]" />
-                    <div className="space-y-2">
-                      <p className="leading-none">6:00 MM, Daily, 200 ml</p>
-                    </div>
-                  </div>
-                </CardContent>
-              </CustomCard>
-            </div>
-          </div>
-          <DrawerFooter className="mx-auto w-full max-w-sm">
-            <Button className="bg-[#6BD9E2] text-xl h-12">
-              <FaPlus />
-            </Button>
-          </DrawerFooter>
-        </DrawerContent>
-      </Drawer>
+      <WaterPreferences />
     </div>
   )
 }
